sitenews/requirements.txt
asgiref==3.4.1
awesome-slugify==1.6.5
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-grappelli==2.15.1
django-ipware==3.0.2
django-js-asset==1.2.2
django-meta==2.0.0
django-mptt==0.12.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
Unidecode==0.4.21
